fix: :recycle: Fix template message
Change template message with list message
