+++ src/commands/general/menu.ts
@@ -1,8 +1,9 @@
+import type { proto } from "@adiwajshing/baileys"
 import os from "os"
 import i18n from "../../libs/international"
 import config from "../../utils/config"
 import { command } from "../../libs/whatsapp"
-import { sizeFormat, timeFormat } from "../../utils/format"
+import { sizeFormat, timeFormat, upperFormat } from "../../utils/format"
 import type { Command } from "../../types/command"
 
 export default <Command>{
@@ -10,9 +11,61 @@ export default <Command>{
   category: "general",
   desc: "Landing menu",
   maintenance: false,
-  execute: async ({ aruga, message, prefix, user }) => {
+  execute: async ({ aruga, message, prefix, user, args, isOwner }) => {
+    if (args.length === 1) {
+      const name = args[0].toLowerCase()
+      const cmd = command.commands.get(name) ?? command.commands.find((cmd) => cmd.aliases && cmd.aliases.includes(name))
+      if ((!cmd || cmd.category === "owner") && !isOwner)
+        return await message.reply(i18n.translate("commands.general.menu.cmd.zero", {}, user.language), true)
+      const text =
+        "┏━━「 𓆩 𝐻ɪᴅᴅᴇɴ 𝐹ɪɴᴅᴇʀ ⁣𓆪 」\n" +
+        "┃\n" +
+        `┃ *${i18n.translate("commands.general.menu.cmd.one", {}, user.language)} :* ${
+          cmd.aliases ? [name].concat(cmd.aliases).join(", ").trim() : name
+        }\n` +
+        `┃ *${i18n.translate("commands.general.menu.cmd.two", {}, user.language)} :* ${cmd.category || "-"}\n` +
+        `┃ *${i18n.translate("commands.general.menu.cmd.three", {}, user.language)} :* ${cmd.desc || "-"}\n` +
+        `┃ *${i18n.translate("commands.general.menu.cmd.four", {}, user.language)} :* ${prefix}${name} ${cmd.example || ``}\n` +
+        `┃ *${i18n.translate("commands.general.menu.cmd.five", {}, user.language)} :* ${cmd.groupOnly ? "✔️" : "✖️"}\n` +
+        (message.isGroupMsg
+          ? `┃ *${i18n.translate("commands.general.menu.cmd.six", {}, user.language)}:* ${cmd.adminGroup ? "✔️" : "✖️"}\n` +
+            `┃ *${i18n.translate("commands.general.menu.cmd.seven", {}, user.language)} :* ${cmd.ownerGroup ? "✔️" : "✖️"}\n`
+          : "") +
+        `┃ *${i18n.translate("commands.general.menu.cmd.eight", {}, user.language)} :* ${cmd.privateOnly ? "✔️" : "✖️"}\n` +
+        `┃ *${i18n.translate("commands.general.menu.cmd.nine", {}, user.language)} :* ${cmd.premiumOnly ? "✔️" : "✖️"}\n` +
+        `┃ *${i18n.translate("commands.general.menu.cmd.ten", {}, user.language)} :* ${cmd.ownerOnly ? "✔️" : "✖️"}\n` +
+        `┃ *${i18n.translate("commands.general.menu.cmd.eleven", {}, user.language)} :* ${cmd.limit ? cmd.limit : "✖️"}\n` +
+        `┃ *${i18n.translate("commands.general.menu.cmd.twelve", {}, user.language)} :* ${cmd.cd ? (cmd.cd % 1000) + "s" : "3s"}\n` +
+        `┃ *${i18n.translate("commands.general.menu.cmd.thirteen", {}, user.language)} :* ${cmd.maintenance ? "✔️" : "✖️"}\n` +
+        "┃\n" +
+        `┗━━「 ꗥ${config.name}ꗥ 」`
+      return await message.reply(text, true)
+    }
+
+    const sections = <proto.Message.ListMessage.ISection[]>[]
+    for (const category of [
+      ...new Set(
+        command.commands
+          .map((v) => v.category)
+          .filter((v) => (!isOwner && v === "owner" ? null : v))
+          .sort()
+      )
+    ]) {
+      const cmd = command.commands.map((v) => v.category === category && v).filter((v) => (!isOwner && v.ownerOnly ? null : v))
+      sections.push({
+        title: upperFormat(category),
+        rows: cmd.map((value) => {
+          const title = command.commands.findKey((v) => v === value)
+          return {
+            title,
+            rowId: `${prefix}menu ${title}`,
+            description: value.desc
+          }
+        })
+      })
+    }
+
     const text =
-      "\n\n" +
       "┏━━「 𓆩 𝐻ɪᴅᴅᴇɴ 𝐹ɪɴᴅᴇʀ ⁣𓆪 」\n" +
       "┃\n" +
       `┃ ${i18n.translate("commands.general.menu.intro.one", { "@PUSHNAME": message.pushname }, user.language)}\n` +
@@ -35,36 +88,17 @@ export default <Command>{
       `┃ ${i18n.translate("commands.general.menu.info.two", {}, user.language)}\n` +
       `┃ ${i18n.translate("commands.general.menu.info.three", { "@COMMANDS": `${prefix}language` }, user.language)}\n` +
       "┃\n" +
-      `┗━━「 ꗥ${config.name}ꗥ 」` +
-      "\n\n"
+      "┣━━━━━━━━━━━━━━━━━━\n" +
+      "┃\n" +
+      `┃ ${i18n.translate("commands.general.menu.bottom", {}, user.language)}\n` +
+      "┃\n" +
+      `┗━━「 ꗥ${config.name}ꗥ 」`
+
     return await aruga.sendMessage(message.from, {
       text,
       footer: config.footer,
-      templateButtons: [
-        {
-          index: 1,
-          urlButton: {
-            displayText: i18n.translate("commands.general.menu.template.one", {}, user.language),
-            url: "https://github.com/arugaz/whatsapp-bot"
-          }
-        },
-        {
-          index: 2,
-          quickReplyButton: {
-            displayText: i18n.translate("commands.general.menu.template.two", {}, user.language),
-            id: `${prefix}listmenu`
-          }
-        },
-        !message.isGroupMsg
-          ? {
-              index: 3,
-              quickReplyButton: {
-                displayText: i18n.translate("commands.general.menu.template.three", {}, user.language),
-                id: `${prefix}language`
-              }
-            }
-          : {}
-      ],
+      sections,
+      buttonText: i18n.translate("commands.general.menu.buttonText", {}, user.language),
       viewOnce: true
     })
   }
